Начал делать адаптив. Я оказывается настолько давно его не делал, что уже и забыл как он делается в том смысле, что не замечаю даже самых простых ошибок, которые впоследствии приходится исправлять. Вот что значит, долго что-то не делать.
