+++ components/ApplicationItem/ApplicationItem.module.scss
