+++ components/ApplicationItem/ApplicationItem.props.ts
@@ -1,7 +1,7 @@
 import { ITag } from "../../interfaces";
 import { IDapplets } from "../../interfaces/redux.state";
 
-export interface DappletsProps {
-	dapplets: IDapplets[];
-	tags: ITag[]
+export interface ApplicationItemProps {
+	dapplets?: IDapplets
+	tags: ITag[];
 }
