+++ components/Search/Search.module.scss
@@ -1,0 +1,34 @@
+.wrapper {
+	position: relative;
+}
+
+.icon {
+	position: absolute;
+	top: 50%;
+	left: 15px;
+
+	width: 16px;
+	height: 16px;
+
+	transform: translateY(-50%);
+}
+
+.input {
+	width: 600px;
+	height: 40px;
+	padding: 0 40px;
+
+	color: #BBB;
+
+	border: none;
+	border-radius: 10px;
+	background: linear-gradient(180deg, rgba(255, 255, 255, 0.2) 0%, rgba(255, 255, 255, 0) 100%), rgba(255, 255, 255, 0.8);
+	box-shadow: 10px 10px 40px rgba(156, 185, 222, 0.36);
+
+	font-size: 16px;
+
+	&:focus {
+		outline: none;
+		box-shadow: 0 0 4px 0 #0085FF;
+	}
+}
