+++ components/Search/Search.props.ts
@@ -1,0 +1,6 @@
+import React from "react";
+
+export interface SearchProps {
+	value?: string;
+	onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void
+}
