+++ components/Select/Select.props.ts
@@ -1,0 +1,5 @@
+export interface SelectProps {
+	values: string[],
+	defaultValue: string;
+	open?: boolean
+}
