+++ layouts/Header/Header.module.scss
@@ -42,3 +42,9 @@
 		margin-left: 10px;
 	}
 }
+
+@media screen and (max-width: 1220px) {
+	.settings {
+		display: none;
+	}
+}
